refactor(storage): move files to corresponding dirs
