+++ cmd/storage/frontend/virtio_blk.go
@@ -1,13 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright (C) 2024 Intel Corporation
 
-// Package cmd implements the CLI commands
-package cmd
+// Package frontend implements the CLI commands for storage frontend
+package frontend
 
 import (
 	"context"
 
-	"github.com/opiproject/godpu/storage"
+	"github.com/opiproject/godpu/cmd/storage/common"
+	frontendclient "github.com/opiproject/godpu/storage/frontend"
 	"github.com/spf13/cobra"
 )
 
@@ -24,13 +25,13 @@ func newCreateVirtioBlkCommand() *cobra.Command {
 		Short:   "Creates virtio-blk controller",
 		Args:    cobra.NoArgs,
 		Run: func(c *cobra.Command, args []string) {
-			addr, err := c.Flags().GetString(addrCmdLineArg)
+			addr, err := c.Flags().GetString(common.AddrCmdLineArg)
 			cobra.CheckErr(err)
 
-			timeout, err := c.Flags().GetDuration(timeoutCmdLineArg)
+			timeout, err := c.Flags().GetDuration(common.TimeoutCmdLineArg)
 			cobra.CheckErr(err)
 
-			client, err := storage.New(addr)
+			client, err := frontendclient.New(addr)
 			cobra.CheckErr(err)
 
 			ctx, cancel := context.WithTimeout(context.Background(), timeout)
@@ -39,7 +40,7 @@ func newCreateVirtioBlkCommand() *cobra.Command {
 			response, err := client.CreateVirtioBlk(ctx, id, volume, port, pf, vf, maxIoQPS)
 			cobra.CheckErr(err)
 
-			printResponse(response.Name)
+			common.PrintResponse(response.Name)
 		},
 	}
 
@@ -66,13 +67,13 @@ func newDeleteVirtioBlkCommand() *cobra.Command {
 		Short:   "Deletes virtio-blk controller",
 		Args:    cobra.NoArgs,
 		Run: func(c *cobra.Command, args []string) {
-			addr, err := c.Flags().GetString(addrCmdLineArg)
+			addr, err := c.Flags().GetString(common.AddrCmdLineArg)
 			cobra.CheckErr(err)
 
-			timeout, err := c.Flags().GetDuration(timeoutCmdLineArg)
+			timeout, err := c.Flags().GetDuration(common.TimeoutCmdLineArg)
 			cobra.CheckErr(err)
 
-			client, err := storage.New(addr)
+			client, err := frontendclient.New(addr)
 			cobra.CheckErr(err)
 
 			ctx, cancel := context.WithTimeout(context.Background(), timeout)
