+++ main.go
@@ -9,6 +9,7 @@ import (
 	"os"
 
 	"github.com/opiproject/godpu/cmd"
+	"github.com/opiproject/godpu/cmd/storage"
 	"github.com/spf13/cobra"
 )
 
@@ -37,7 +38,7 @@ func newCommand() *cobra.Command {
 	}
 	c.AddCommand(cmd.NewInventoryCommand())
 	c.AddCommand(cmd.NewIPSecCommand())
-	c.AddCommand(cmd.NewStorageCommand())
+	c.AddCommand(storage.NewStorageCommand())
 	c.AddCommand(cmd.NewEvpnCommand())
 	return c
 }
